// enums/docusign-text-tabs.ts
export enum DOCUSIGN_TEXT_TABS {
  CUSTOMER_NAME = "CUSTOMER_NAME",
  CUSTOMER_PHONE = "CUSTOMER_PHONE",
  CUSTOMER_EMAIL = "CUSTOMER_EMAIL",
  JOB_FULL_ADDRESS = "JOB_FULL_ADDRESS",
  JOB_ADDRESS = "JOB_ADDRESS",
  JOB_CITY = "JOB_CITY",
  JOB_STATE = "JOB_STATE",
  JOB_POSTAL_CODE = "JOB_POSTAL_CODE",
  CREATOR_NAME = "CREATOR_NAME",
  CREATOR_EMAIL = "CREATOR_EMAIL",
}